build(api): internal patient endpoints get cxId from query
- internal patient endpoints get cxId from query
- move delete patient/:id to internal patient routes
- remove duplicated delete patient endpoint

Ref. metriport/metriport-internal#715
